[S14P11E206-424] feat(be): 제출하기 최신 결과 조회 구현

+++ backend/src/main/java/io/ssafy/trycatch/domain/submission/dto/response/SubmissionRespDto.java
@@ -14,7 +14,7 @@ public class SubmissionRespDto {
     private Long submissionId;
     private Long roomId;
     private Long questId;
-    private Long questOrder;
+    private Integer questOrder;
     private String status;
     private Integer score;
     private Long executionTimeMs;
